feat: reuse ModalWithToggler for AppSelector
