+++ webapp/javascript/components/AppSelector/index.tsx
@@ -1,4 +1,6 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
+import { faSyncAlt } from '@fortawesome/free-solid-svg-icons/faSyncAlt';
+
 import { queryFromAppName, queryToAppName, Query } from '@webapp/models/query';
 import { useAppSelector, useAppDispatch } from '@webapp/redux/hooks';
 import {
@@ -10,39 +12,23 @@ import {
 } from '@webapp/redux/reducers/continuous';
 import Button from '@webapp/ui/Button';
 import LoadingSpinner from '@webapp/ui/LoadingSpinner';
-import OutsideClickHandler from 'react-outside-click-handler';
-import { faSyncAlt } from '@fortawesome/free-solid-svg-icons/faSyncAlt';
-import SelectorModal from './SelectorModal';
+import ModalWithToggle from '@webapp/ui/Modals/ModalWithToggle';
+import Input from '@webapp/ui/Input';
+import SelectButton from './SelectButton';
 import styles from './AppSelector.module.scss';
 
 interface AppSelectorProps {
   onSelectedName?: (name: Query) => void;
 }
 
-export const TOGGLE_BTN_ID = 'app_selector_toggle_button';
-
-const Loading = ({ type }: { type: 'reloading' | 'loaded' | 'failed' }) => {
-  switch (type) {
-    case 'reloading': {
-      return <LoadingSpinner />;
-    }
-
-    default: {
-      return null;
-    }
-  }
-};
-
 const AppSelector = ({ onSelectedName }: AppSelectorProps) => {
   const dispatch = useAppDispatch();
+  const appNamesState = useAppSelector(selectAppNamesState);
   const appNames = useAppSelector(selectAppNames);
   const { query } = useAppSelector(selectQueries);
   const appName = queryToAppName(query).mapOr('', (q) =>
     appNames.indexOf(q) !== -1 ? q : ''
   );
-  const appNamesState = useAppSelector(selectAppNamesState);
-
-  const [selectorOpened, toggleSelector] = useState(false);
 
   const selectAppName = (name: string) => {
     const appNameQuery = queryFromAppName(name);
@@ -53,45 +39,188 @@ const AppSelector = ({ onSelectedName }: AppSelectorProps) => {
     } else {
       dispatch(actions.setQuery(appNameQuery));
     }
-
-    toggleSelector(false);
-  };
-
-  const handleClickOutsile = (e: MouseEvent) => {
-    if ((e.target as { id?: string })?.id !== TOGGLE_BTN_ID) {
-      toggleSelector(false);
-    }
   };
 
   return (
     <div className={styles.container}>
       Application:&nbsp;
-      <button
-        id={TOGGLE_BTN_ID}
-        data-testid={TOGGLE_BTN_ID}
-        className={styles.toggleButton}
-        onClick={() => toggleSelector(!selectorOpened)}
-        type="button"
-      >
-        {appName || 'Select application'}
-      </button>
+      <SelectorModalWithToggler
+        selectAppName={selectAppName}
+        appNames={appNames}
+        appName={appName}
+      />
       <Button
         aria-label="Refresh Apps"
         icon={faSyncAlt}
         onClick={() => dispatch(reloadAppNames())}
         className={styles.refreshButton}
       />
-      <Loading type={appNamesState.type} />
-      <OutsideClickHandler onOutsideClick={handleClickOutsile}>
-        <SelectorModal
-          selectAppName={selectAppName}
-          appNames={appNames}
-          visible={selectorOpened}
-          appName={appName}
-        />
-      </OutsideClickHandler>
+      {appNamesState.type === 'reloading' && <LoadingSpinner />}
     </div>
   );
 };
 
 export default AppSelector;
+
+const DELIMITER = '.';
+const isGroupMember = (groupName: string, name: string) =>
+  name.indexOf(groupName) === 0 &&
+  (name[groupName.length] === DELIMITER || name.length === groupName.length);
+
+const getGroupMembers = (names: string[], name: string) =>
+  names.filter((a) => isGroupMember(name, a));
+
+const getGroupNameFromAppName = (groups: string[], fullName: string) =>
+  groups.filter((g) => isGroupMember(g, fullName))[0] || '';
+
+const getGroups = (filteredAppNames: string[]) => {
+  const allGroups = filteredAppNames.map((i) => {
+    const arr = i.split(DELIMITER);
+    const cutProfileType = arr.length > 1 ? arr.slice(0, -1) : arr;
+    return cutProfileType.join(DELIMITER);
+  });
+
+  const uniqGroups = Array.from(new Set(allGroups));
+
+  const dedupedUniqGroups = uniqGroups.filter((x) => {
+    return !uniqGroups.find((y) => x !== y && isGroupMember(y, x));
+  });
+
+  const groupOrApp = dedupedUniqGroups.map((u) => {
+    const appNamesEntries = getGroupMembers(filteredAppNames, u);
+
+    return appNamesEntries.length > 1 ? u : appNamesEntries?.[0];
+  });
+
+  return groupOrApp;
+};
+
+interface SelectorModalWithTogglerProps {
+  appNames: string[];
+  selectAppName: (name: string) => void;
+  appName: string;
+}
+
+const SelectorModalWithToggler = ({
+  appNames,
+  selectAppName,
+  appName,
+}: SelectorModalWithTogglerProps) => {
+  const [filter, setFilter] = useState('');
+
+  // selected is an array of strings
+  //  0 corresponds to string of group / app name selected in the left pane
+  //  1 corresponds to string of app name selected in the right pane
+  const [selected, setSelected] = useState<string[]>([]);
+  const filteredAppNames = useMemo(
+    // filtered names by search input
+    () =>
+      appNames.filter((n: string) =>
+        n.toLowerCase().includes(filter.trim().toLowerCase())
+      ),
+    [filter, appNames]
+  );
+
+  const groups = useMemo(() => getGroups(filteredAppNames), [filteredAppNames]);
+
+  const profileTypes = useMemo(() => {
+    if (!selected?.[0]) {
+      return [];
+    }
+
+    const filtered = getGroupMembers(filteredAppNames, selected?.[0]);
+
+    if (filtered.length > 1) {
+      return filtered;
+    }
+
+    return [];
+  }, [selected, groups, filteredAppNames]);
+
+  const onSelect = ({ index, name }: { index: number; name: string }) => {
+    const filtered = getGroupMembers(filteredAppNames, name);
+
+    if (filtered.length === 1 || index === 1) {
+      selectAppName(filtered?.[0]);
+    }
+
+    const arr = Array.from(selected);
+
+    if (index === 0 && arr?.length > 1) {
+      arr.pop();
+    }
+
+    arr[index] = name;
+
+    setSelected(arr);
+  };
+
+  useEffect(() => {
+    if (appName && !selected.length && groups.length) {
+      if (groups.indexOf(appName) !== -1) {
+        setSelected([appName]);
+      } else {
+        setSelected([getGroupNameFromAppName(groups, appName), appName]);
+      }
+    }
+  }, [appName, selected, groups]);
+
+  const listHeight = useMemo(() => {
+    const height = (window?.innerHeight || 0) - 160;
+
+    const listRequiredHeight =
+      // 35 is list item height
+      Math.max(groups?.length || 0, profileTypes?.length || 0) * 35;
+
+    if (height && listRequiredHeight) {
+      return height >= listRequiredHeight ? 'auto' : `${height}px`;
+    }
+
+    return 'auto';
+  }, [groups, profileTypes]);
+
+  return (
+    <ModalWithToggle
+      modalClassName={styles.appSelectorModal}
+      modalHeight={listHeight}
+      noDataEl={
+        filteredAppNames?.length ? (
+          <div className={styles.noData}>No Data</div>
+        ) : null
+      }
+      toggleText={appName || 'Select application'}
+      headerEl={
+        <>
+          <div className={styles.headerTitle}>SELECT APPLICATION</div>
+          <Input
+            name="application seach"
+            type="text"
+            placeholder="Type an app"
+            value={filter}
+            onChange={(e) => setFilter(e.target.value)}
+            className={styles.search}
+            testId="application-search"
+          />
+        </>
+      }
+      leftSideEl={groups.map((name) => (
+        <SelectButton
+          name={name}
+          onClick={() => onSelect({ index: 0, name })}
+          fullList={appNames}
+          isSelected={selected?.[0] === name}
+          key={name}
+        />
+      ))}
+      rightSideEl={profileTypes.map((name) => (
+        <SelectButton
+          name={name}
+          onClick={() => onSelect({ index: 1, name })}
+          fullList={appNames}
+          isSelected={selected?.[1] === name}
+          key={name}
+        />
+      ))}
+    />
+  );
+};
