+++ webapp/javascript/ui/Modals/ModalWithToggle.module.scss
@@ -28,6 +28,7 @@
   background-color: var(--ps-dropdown-background);
   border: 1px solid var(--ps-ui-border);
   border-radius: 4px;
+  box-shadow: 0px 2px 6px var(--ps-dropdown-shadow);
 
   .modalHeader {
     padding: 12px 10px;
@@ -40,6 +41,7 @@
     .side {
       width: 50%;
       text-align: center;
+      overflow-y: auto;
 
       &:first-child {
         border-right: 1px solid var(--ps-ui-border);
